+++ packages/abstraxion/src/hooks/useAbstraxionAccount.ts
@@ -1,8 +1,9 @@
-import { useContext } from "react";
+import { useContext, useEffect } from "react";
 import {
   AbstraxionContext,
   AbstraxionContextProps,
 } from "@/src/components/AbstraxionContext";
+import { DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
 
 export interface AbstraxionAccount {
   bech32Address: string;
@@ -14,13 +15,51 @@ export interface useAbstraxionAccountProps {
 }
 
 export const useAbstraxionAccount = (): useAbstraxionAccountProps => {
-  const { isConnected, grantorAddress } = useContext(
-    AbstraxionContext,
-  ) as AbstraxionContextProps;
+  const {
+    isConnected,
+    granterAddress,
+    abstraxionAccount,
+    isConnecting,
+    setgranterAddress,
+    setAbstraxionAccount,
+    setIsConnected,
+    setIsConnecting,
+  } = useContext(AbstraxionContext) as AbstraxionContextProps;
+
+  useEffect(() => {
+    async function configureAccount() {
+      setIsConnecting(true);
+      const tempKeypair = localStorage.getItem("xion-authz-temp-account");
+      if (tempKeypair) {
+        const deserializedKeypair = await DirectSecp256k1HdWallet.deserialize(
+          tempKeypair,
+          "abstraxion",
+        );
+        setAbstraxionAccount(deserializedKeypair);
+        const granterAccount = localStorage.getItem(
+          "xion-authz-granter-account",
+        );
+        if (granterAccount) {
+          setgranterAddress(granterAccount);
+          setIsConnected(true);
+        }
+      } else {
+        // Wipe granter even if it exists, clean context
+        localStorage.removeItem("xion-authz-granter-account");
+        setAbstraxionAccount(undefined);
+        setgranterAddress("");
+      }
+      setIsConnecting(false);
+    }
+
+    if (!isConnecting && !abstraxionAccount && !granterAddress) {
+      configureAccount();
+    }
+  }, [isConnected, abstraxionAccount, granterAddress]);
 
   return {
     data: {
-      bech32Address: grantorAddress,
+      bech32Address: granterAddress,
     },
     isConnected: isConnected,
   };